Split `routers/recap.py` based on prefix
I decided to split the `recap.py` router up so users can swap them in and out
independently. A user might want to use the `directory` router, but might have a
much more specific metadata router they'd like to use (that uses Pydantic types,
for example). Splitting up the routers allows for this.
